pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "A collection of utility scripts"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "pillow>=11.0",
    "beautifulsoup4>=4.14.3",
    "markdownify>=1.2.2",
    "lxml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff",
    "mypy",
]
